ping-py-bot/requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
aiohttp==3.6.2            # via discord.py
async-timeout==3.0.1      # via aiohttp
attrs==20.1.0             # via aiohttp
chardet==3.0.4            # via aiohttp
discord.py==1.4.1         # via discord
discord==1.0.1            # via -r requirements/requirements.in
idna==2.10                # via yarl
multidict==4.7.6          # via aiohttp, yarl
python-dotenv==0.14.0     # via -r requirements/requirements.in
yarl==1.5.1               # via aiohttp
